+++ sentry-log4j2/src/main/java/io/sentry/log4j2/SentryAppender.java
@@ -4,13 +4,13 @@ import io.sentry.Breadcrumb;
 import io.sentry.DateUtils;
 import io.sentry.HubAdapter;
 import io.sentry.IHub;
+import io.sentry.ITransportFactory;
 import io.sentry.Sentry;
 import io.sentry.SentryEvent;
 import io.sentry.SentryLevel;
 import io.sentry.SentryOptions;
 import io.sentry.protocol.Message;
 import io.sentry.protocol.SdkVersion;
-import io.sentry.transport.ITransport;
 import io.sentry.util.CollectionUtils;
 import java.util.Arrays;
 import java.util.Collections;
@@ -35,7 +35,7 @@ import org.jetbrains.annotations.Nullable;
 @Plugin(name = "Sentry", category = "Core", elementType = "appender", printObject = true)
 public final class SentryAppender extends AbstractAppender {
   private final @Nullable String dsn;
-  private final @Nullable ITransport transport;
+  private final @Nullable ITransportFactory transportFactory;
   private @NotNull Level minimumBreadcrumbLevel = Level.INFO;
   private @NotNull Level minimumEventLevel = Level.ERROR;
   private final @NotNull IHub hub;
@@ -46,7 +46,7 @@ public final class SentryAppender extends AbstractAppender {
       final @Nullable String dsn,
       final @Nullable Level minimumBreadcrumbLevel,
       final @Nullable Level minimumEventLevel,
-      final @Nullable ITransport transport,
+      final @Nullable ITransportFactory transportFactory,
       final @NotNull IHub hub) {
     super(name, filter, null, true, null);
     this.dsn = dsn;
@@ -56,7 +56,7 @@ public final class SentryAppender extends AbstractAppender {
     if (minimumEventLevel != null) {
       this.minimumEventLevel = minimumEventLevel;
     }
-    this.transport = transport;
+    this.transportFactory = transportFactory;
     this.hub = hub;
   }
 
@@ -102,7 +102,7 @@ public final class SentryAppender extends AbstractAppender {
               options.setDsn(dsn);
               options.setSentryClientName(BuildConfig.SENTRY_LOG4J2_SDK_NAME);
               options.setSdkVersion(createSdkVersion(options));
-              Optional.ofNullable(transport).ifPresent(options::setTransport);
+              Optional.ofNullable(transportFactory).ifPresent(options::setTransportFactory);
             });
       } catch (IllegalArgumentException e) {
         LOGGER.info("Failed to init Sentry during appender initialization: " + e.getMessage());