+++ sentry-spring-boot-starter/src/test/kotlin/io/sentry/spring/boot/SentryAutoConfigurationTest.kt
@@ -1,13 +1,18 @@
 package io.sentry.spring.boot
 
 import com.acme.MainBootClass
+import com.nhaarman.mockitokotlin2.any
+import com.nhaarman.mockitokotlin2.anyOrNull
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
+import io.sentry.AsyncHttpTransportFactory
 import io.sentry.Breadcrumb
 import io.sentry.EventProcessor
 import io.sentry.IHub
+import io.sentry.ITransportFactory
 import io.sentry.Integration
+import io.sentry.NoOpTransportFactory
 import io.sentry.SamplingContext
 import io.sentry.Sentry
 import io.sentry.SentryEvent
@@ -21,6 +26,7 @@ import io.sentry.spring.tracing.SentryTracingFilter
 import io.sentry.test.checkEvent
 import io.sentry.transport.ITransport
 import io.sentry.transport.ITransportGate
+import io.sentry.transport.apache.ApacheHttpClientTransportFactory
 import javax.servlet.Filter
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -186,7 +192,7 @@ class SentryAutoConfigurationTest {
                         assertThat(event.sdk.packages).anyMatch { pkg ->
                             pkg.name == "maven:sentry-spring-boot-starter" && pkg.version == BuildConfig.VERSION_NAME
                         }
-                    })
+                    }, anyOrNull())
                 }
             }
     }
@@ -246,7 +252,7 @@ class SentryAutoConfigurationTest {
                 await.untilAsserted {
                     verify(transport).send(checkEvent { event ->
                         assertThat(event.release).isEqualTo("git-commit-id")
-                    })
+                    }, anyOrNull())
                 }
             }
     }
@@ -261,7 +267,7 @@ class SentryAutoConfigurationTest {
                 await.untilAsserted {
                     verify(transport).send(checkEvent { event ->
                         assertThat(event.release).isEqualTo("my-release")
-                    })
+                    }, anyOrNull())
                 }
             }
     }
@@ -452,6 +458,34 @@ class SentryAutoConfigurationTest {
             }
     }
 
+    @Test
+    fun `when sentry-apache-http-client-5 is on the classpath, creates apache transport factory`() {
+        contextRunner.withPropertyValues("sentry.dsn=http://key@localhost/proj")
+            .run {
+                assertThat(it.getBean(SentryOptions::class.java).transportFactory).isInstanceOf(ApacheHttpClientTransportFactory::class.java)
+            }
+    }
+
+    @Test
+    fun `when sentry-apache-http-client-5 is not on the classpath, does not create apache transport factory`() {
+        contextRunner.withPropertyValues("sentry.dsn=http://key@localhost/proj")
+            .withClassLoader(FilteredClassLoader(ApacheHttpClientTransportFactory::class.java))
+            .run {
+                assertThat(it.getBean(SentryOptions::class.java).transportFactory).isInstanceOf(AsyncHttpTransportFactory::class.java)
+            }
+    }
+
+    @Test
+    fun `when sentry-apache-http-client-5 is on the classpath and custom transport factory bean is set, does not create apache transport factory`() {
+        contextRunner.withPropertyValues("sentry.dsn=http://key@localhost/proj")
+            .withUserConfiguration(MockTransportConfiguration::class.java)
+            .run {
+                assertThat(it.getBean(SentryOptions::class.java).transportFactory)
+                    .isNotInstanceOf(ApacheHttpClientTransportFactory::class.java)
+                    .isNotInstanceOf(NoOpTransportFactory::class.java)
+            }
+    }
+
     @Configuration(proxyBeanMethods = false)
     open class CustomOptionsConfigurationConfiguration {
 
@@ -475,8 +509,17 @@ class SentryAutoConfigurationTest {
     @Configuration(proxyBeanMethods = false)
     open class MockTransportConfiguration {
 
+        private val transport = mock<ITransport>()
+
         @Bean
-        open fun sentryTransport() = mock<ITransport>()
+        open fun mockTransportFactory(): ITransportFactory {
+            val factory = mock<ITransportFactory>()
+            whenever(factory.create(any(), any())).thenReturn(transport)
+            return factory
+        }
+
+        @Bean
+        open fun sentryTransport() = transport
     }
 
     @Configuration(proxyBeanMethods = false)