+++ sentry/src/main/java/io/sentry/Hub.java
@@ -3,6 +3,7 @@ package io.sentry;
 import io.sentry.Stack.StackItem;
 import io.sentry.hints.SessionEndHint;
 import io.sentry.hints.SessionStartHint;
+import io.sentry.protocol.Contexts;
 import io.sentry.protocol.SentryId;
 import io.sentry.protocol.User;
 import io.sentry.util.Objects;
@@ -78,6 +79,7 @@ public final class Hub implements IHub {
       options.getLogger().log(SentryLevel.WARNING, "captureEvent called with null parameter.");
     } else {
       try {
+        assignTraceContext(event);
         final StackItem item = stack.peek();
         sentryId = item.getClient().captureEvent(event, item.getScope(), hint);
       } catch (Exception e) {
@@ -154,7 +156,9 @@ public final class Hub implements IHub {
     } else {
       try {
         final StackItem item = stack.peek();
-        sentryId = item.getClient().captureException(throwable, item.getScope(), hint);
+        final SentryEvent event = new SentryEvent(throwable);
+        assignTraceContext(event);
+        sentryId = item.getClient().captureEvent(event, item.getScope(), hint);
       } catch (Exception e) {
         options
             .getLogger()
@@ -164,6 +168,20 @@ public final class Hub implements IHub {
     }
     this.lastEventId = sentryId;
     return sentryId;
+  }
+
+  private void assignTraceContext(final @NotNull SentryEvent event) {
+    if (event.getThrowable() != null) {
+      final SpanContext spanContext = throwableToSpanContext.get(event.getThrowable());
+      if (spanContext != null) {
+        if (event.getContexts() == null) {
+          event.setContexts(new Contexts());
+          event.getContexts().setTrace(spanContext);
+        } else if (event.getContexts().getTrace() == null) {
+          event.getContexts().setTrace(spanContext);
+        }
+      }
+    }
   }
 
   @Override