+++ sentry/src/main/java/io/sentry/transport/ITransport.java
@@ -6,7 +6,9 @@ import java.io.IOException;
 
 /** A transport is in charge of sending the event to the Sentry server. */
 public interface ITransport extends Closeable {
-  boolean isRetryAfter(String type);
+  void send(SentryEnvelope envelope, Object hint) throws IOException;
 
-  TransportResult send(SentryEnvelope envelope) throws IOException;
+  default void send(SentryEnvelope envelope) throws IOException {
+    send(envelope, null);
+  }
 }